linux-dist: Move most x11 related code into a separate file
