+++ core/linux-dist/main.h
@@ -1,9 +1,13 @@
 #pragma once
+#include "types.h"
 
 extern u16 kcode[4];
 extern u32 vks[4];
 extern u8 rt[4], lt[4];
 extern s8 joyx[4], joyy[4];
+
+extern void* x11_win;
+extern void* x11_disp;
 
 enum DreamcastController
 {
