+++ core/linux-dist/x11.h
@@ -1,0 +1,7 @@
+#pragma once
+
+extern void* x11_glc;
+extern void input_x11_init();
+extern void input_x11_handle();
+extern void x11_window_create();
+extern void x11_window_set_text(const char* text);
